+++ src/main/java/org/kumoricon/site/report/ReportView.java
@@ -5,4 +5,5 @@ import org.kumoricon.model.user.User;
 public interface ReportView {
     void afterSuccessfulFetch(String data);
     User getCurrentUser();
+    Boolean currentUserHasRight(String right);
 }