+++ CHANGES.md
@@ -7,6 +7,7 @@ Change Log
     - Use computer specific offsets when displaying badge PDFs in browser
     - Added base data for one-day events
     - Removed unnecessary error logging from BadgePresenter
+    - Added revenue to Attendance report with right view_attendance_report_revenue
 
 - Version 0.9.3 (8/6/2016)
     - Select the number of badges to generate when printing test badges